requirements.txt
absl-py==2.1.0
accelerate==1.0.1
addict==2.4.0
aiohttp==3.9.1
aiosignal==1.3.1
alembic==1.13.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
antlr4-python3-runtime==4.9.3
async-timeout==4.0.3
attrs==23.1.0
Automat==24.8.1
blinker==1.7.0
buildtools==1.0.6
calflops==0.3.2
causal-conv1d==1.1.0
certifi==2023.11.17
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
constantly==23.10.4
contourpy==1.2.0
crcmod==1.7
cryptography==44.0.0
cycler==0.12.1
databricks-cli==0.18.0
datasets==2.15.0
decord==0.6.0
dill==0.3.7
docker==6.1.3
docopt==0.6.2
einops==0.7.0
entrypoints==0.4
filelock==3.14.0
Flask==3.0.0
fonttools==4.46.0
frozenlist==1.4.0
fsspec==2023.10.0
furl==2.1.3
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.2
grpcio==1.67.0
gunicorn==21.2.0
h5py==3.12.1
huggingface-hub==0.25.2
hyperlink==21.0.0
idna==3.6
importlib-metadata==7.0.0
incremental==24.7.2
iopath==0.1.10
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.3.2
kiwisolver==1.4.5
Mako==1.3.0
-e git+https://github.com/hustvl/Vim.git@1be434d49d81f24c7d1aef3f4729d6e64a1c3c62#egg=mamba_ssm&subdirectory=mamba-1p1p1
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
mlflow==2.9.1
mmcv==1.3.8
mmsegmentation==0.14.1
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.2
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.8.1.78
openxlab==0.1.2
orderedmultidict==1.0.1
oss2==2.17.0
pandas==2.1.3
Pillow==10.1.0
platformdirs==4.1.0
portalocker==2.10.1
prettytable==3.9.0
protobuf==4.25.1
psutil==6.0.0
ptflops==0.7.4
pyarrow==14.0.1
pyarrow-hotfix==0.6
pycparser==2.22
pycryptodome==3.21.0
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-hostlist==1.23.0
pytz==2023.3.post1
PyYAML==6.0.1
querystring-parser==1.2.4
redo==3.0.0
regex==2023.10.3
requests==2.28.2
rich==13.4.2
safetensors==0.4.5
scikit-learn==1.3.2
scipy==1.11.4
simplejson==3.19.3
smmap==5.0.1
SQLAlchemy==2.0.23
sqlparse==0.4.4
sympy==1.12
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.5.0
threadpoolctl==3.2.0
timm==0.4.12
tokenizers==0.15.0
tomli==2.0.1
torch==2.1.1+cu118
torchaudio==2.1.1+cu118
torchvision==0.16.1+cu118
tqdm==4.65.2
transformers==4.35.2
triton==2.1.0
Twisted==24.7.0
tzdata==2023.3
u==0.1
urllib3==1.26.20
wcwidth==0.2.12
websocket-client==1.7.0
Werkzeug==3.0.1
xxhash==3.4.1
yacs==0.1.8
yapf==0.40.2
yarl==1.9.4
zipp==3.17.0
zope.interface==7.1.0
